pyproject.toml
[tool.poetry]
name = "projeto02"
version = "0.1.0"
description = "Projeto 2 curso de FastAPI TDD."
authors = ["Reinaldo Silva <josereinaldo.pessoal@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["all"], version = "^0.111.0" }
sqlalchemy = "^2.0.30"
pytest = "^8.2.1"
alembic = "^1.13.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
python-jose = "^3.3.0"
uvicorn = "^0.29.0"
fastapi-pagination = "^0.12.24"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"


[tool.black]
line-length = 88
target-version = ['py310']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
